+++ src/core/taskjudger.cpp
@@ -361,7 +361,7 @@ int TaskJudger::judge() {
 
 	for (int i = 0; i < task->getTestCaseList().size(); i++) {
 		timeUsed.append(QList<int>());
-		memoryUsed.append(QList<int>());
+		memoryUsed.append(QList<long long>());
 		score.append(QList<int>());
 		result.append(QList<ResultState>());
 		overallStatus.append(maxDependValue);
